requirements.txt
openstacksdk
tabulate
colorama